requirements.txt
# CodePromptZip Requirements
# GPU CUDA 12.6 Torch Wheel Index
--extra-index-url https://download.pytorch.org/whl/cu126

# Core ML
torch>=2.6.0
torchvision
torchaudio
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
safetensors

# Code analysis
javalang>=0.13.0

# Retrieval
rank-bm25>=0.2.2

# Metrics
codebleu>=0.7.0
tree-sitter==0.22.3
tree-sitter-java==0.23.5
nltk>=3.8.0

# Config & Utilities
pyyaml>=6.0
tqdm>=4.66.0
numpy>=1.24.0
pandas>=1.5.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM inference (CodeLlama via llama-cpp)
# NOTE: llama-cpp-python requires C++ build tools.
# Install separately with a pre-built wheel:
#   pip install llama-cpp-python --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu121
# Or install after setting up Visual Studio Build Tools.

# Logging
tensorboard>=2.14.0
